+++ packages/gcp/src/mistralai/gcp/client/models/completionresponsestreamchoice.py
@@ -44,7 +44,7 @@ class CompletionResponseStreamChoice(BaseModel):
 
         for n, f in type(self).model_fields.items():
             k = f.alias or n
-            val = serialized.get(k)
+            val = serialized.get(k, serialized.get(n))
 
             if val != UNSET_SENTINEL:
                 m[k] = val