+++ src/mistralai/client/judges.py
@@ -101,7 +101,7 @@ class Judges(BaseSDK):
                 ),
             ),
             request=req,
-            error_status_codes=["400", "404", "408", "409", "422", "4XX", "5XX"],
+            is_error_status_code=lambda c: utils.match_status_codes(["4XX", "5XX"], c),
             retry_config=retry_config,
         )
 
@@ -214,7 +214,7 @@ class Judges(BaseSDK):
                 ),
             ),
             request=req,
-            error_status_codes=["400", "404", "408", "409", "422", "4XX", "5XX"],
+            is_error_status_code=lambda c: utils.match_status_codes(["4XX", "5XX"], c),
             retry_config=retry_config,
         )
 
@@ -319,7 +319,7 @@ class Judges(BaseSDK):
                 ),
             ),
             request=req,
-            error_status_codes=["400", "404", "408", "409", "422", "4XX", "5XX"],
+            is_error_status_code=lambda c: utils.match_status_codes(["4XX", "5XX"], c),
             retry_config=retry_config,
         )
 
@@ -424,7 +424,7 @@ class Judges(BaseSDK):
                 ),
             ),
             request=req,
-            error_status_codes=["400", "404", "408", "409", "422", "4XX", "5XX"],
+            is_error_status_code=lambda c: utils.match_status_codes(["4XX", "5XX"], c),
             retry_config=retry_config,
         )
 
@@ -517,7 +517,7 @@ class Judges(BaseSDK):
                 ),
             ),
             request=req,
-            error_status_codes=["400", "404", "408", "409", "422", "4XX", "5XX"],
+            is_error_status_code=lambda c: utils.match_status_codes(["4XX", "5XX"], c),
             retry_config=retry_config,
         )
 
@@ -610,7 +610,7 @@ class Judges(BaseSDK):
                 ),
             ),
             request=req,
-            error_status_codes=["400", "404", "408", "409", "422", "4XX", "5XX"],
+            is_error_status_code=lambda c: utils.match_status_codes(["4XX", "5XX"], c),
             retry_config=retry_config,
         )
 
@@ -703,7 +703,7 @@ class Judges(BaseSDK):
                 ),
             ),
             request=req,
-            error_status_codes=["400", "404", "408", "409", "422", "4XX", "5XX"],
+            is_error_status_code=lambda c: utils.match_status_codes(["4XX", "5XX"], c),
             retry_config=retry_config,
         )
 
@@ -796,7 +796,7 @@ class Judges(BaseSDK):
                 ),
             ),
             request=req,
-            error_status_codes=["400", "404", "408", "409", "422", "4XX", "5XX"],
+            is_error_status_code=lambda c: utils.match_status_codes(["4XX", "5XX"], c),
             retry_config=retry_config,
         )
 
@@ -920,7 +920,7 @@ class Judges(BaseSDK):
                 ),
             ),
             request=req,
-            error_status_codes=["400", "404", "408", "409", "422", "4XX", "5XX"],
+            is_error_status_code=lambda c: utils.match_status_codes(["4XX", "5XX"], c),
             retry_config=retry_config,
         )
 
@@ -1044,7 +1044,7 @@ class Judges(BaseSDK):
                 ),
             ),
             request=req,
-            error_status_codes=["400", "404", "408", "409", "422", "4XX", "5XX"],
+            is_error_status_code=lambda c: utils.match_status_codes(["4XX", "5XX"], c),
             retry_config=retry_config,
         )
 
@@ -1154,7 +1154,7 @@ class Judges(BaseSDK):
                 ),
             ),
             request=req,
-            error_status_codes=["400", "404", "408", "409", "422", "4XX", "5XX"],
+            is_error_status_code=lambda c: utils.match_status_codes(["4XX", "5XX"], c),
             retry_config=retry_config,
         )
 
@@ -1264,7 +1264,7 @@ class Judges(BaseSDK):
                 ),
             ),
             request=req,
-            error_status_codes=["400", "404", "408", "409", "422", "4XX", "5XX"],
+            is_error_status_code=lambda c: utils.match_status_codes(["4XX", "5XX"], c),
             retry_config=retry_config,
         )